优化

+++ Server/Model/Entity/Config/ScoreConfig.cs
@@ -11,8 +11,8 @@ namespace ETModel
     public class ScoreConfig: IConfig
     {
         public long Id { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
     }
 }
